add a test for trace_item_filters_to_expression for logs
